Softserve/UA-4148-alpha-develop/backend/requirements.txt
asgiref==3.9.1
astroid==3.3.11
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==25.1.0
bleach==6.2.0
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
constantly==23.10.4
coverage==7.10.1
cryptography==45.0.6
daphne==4.2.1
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-yasg==1.21.10
environ==1.0
exceptiongroup==1.3.0
flake8==7.3.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mongoengine==0.29.1
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pymongo==4.14.0
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-cov==4.1.0
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
uritemplate==4.2.0
zope.interface==7.2

pytest-django==4.11.1
djangorestframework_simplejwt==5.5.1
coverage==7.10.1
pytest-cov==4.1.0
channels==4.3.1
daphne==4.2.1
channels-redis==4.3.0
Faker==37.5.3
django-filter==2.4.0